* change int values in POJO to long * add catch exception in send and map OrderCreateRequest * create lower range of change price and qty * add commission accounting

+++ src/main/java/me/ikosarim/cripto_bot/json_model/TradeInfoEntity.java
@@ -13,7 +13,7 @@ import java.sql.Timestamp;
 public class TradeInfoEntity {
 
     @JsonProperty(value = "trade_id")
-    private Integer tradeId;
+    private Long tradeId;
     @JsonProperty(value = "type")
     private String type;
     @JsonProperty(value = "quantity")
